refactor: ModuleResolver/execute_imports 移入 vm 层,stmt ↔ vm 循环消除
加载一个模块 = 解析它、运行它、把导出绑进 VmContext——这是执行行为,却
住在 stmt/import.rs 里挨着 use 语句的 AST 类型。后果是 AST 层依赖执行器,
形成 stmt ↔ vm 循环,挡住前端与 VM 的任何分离。

- ModuleResolver、execute_imports、runtime_export_field、
  resolve_runtime_import_source、default_module_binding → 新的
  core/src/vm/resolver.rs;相关测试一并搬(它们测的就是解析/执行)
- stmt/import.rs 757 → 123 行,只剩语法:ImportStmt/ImportSource/
  ImportItem、序列化、collect_program_imports(AST 遍历)
- test_support(run_program*)也移入 vm——它同样是'执行',留在 stmt 会让
  测试构建下的依赖倒置照旧
- 51 处 ModuleResolver 引用改为 lk_core::vm::(未留兼容 re-export:
  留了循环就还在)

结果:**stmt 现在只依赖 compat 和 token**;core 生产代码的循环 5 → 3。
剩余三个(val↔vm、rt↔vm、macro_system↔package)及其窄边记在 CLAUDE.md。

workspace 108 全绿;clippy 零警告;coverage 51/51;
lk-core --no-default-features(no_std)编译通过;strict-native 差分 +
examples + compile_cli 全绿;跨模块 trait 分派仍 = 16
