+++ cli/src/repl.rs
@@ -4,9 +4,9 @@ use std::{
     sync::Arc,
 };
 
+use lk_core::vm::ModuleResolver;
 use lk_core::{
     module::ModuleRegistry,
-    stmt::ModuleResolver,
     syntax::{ParseOptions, parse_program_source},
     typ::TypeChecker,
     vm::{ReplExecutionResult, ReplVmSession, VmContext},
