+++ core/src/stmt.rs
@@ -23,23 +23,3 @@ mod stmt_test;
 pub use import::*;
 pub use stmt_impl::*;
 pub use stmt_parser::*;
-
-#[cfg(test)]
-pub use test_support::*;
-
-#[cfg(test)]
-pub mod test_support {
-    use super::*;
-    use crate::vm::ProgramExec;
-    use crate::vm::{ProgramResult, VmContext};
-    use anyhow::Result;
-
-    pub fn run_program(program: &Program, ctx: &mut VmContext) -> Result<ProgramResult> {
-        program.execute_with_ctx(ctx)
-    }
-
-    pub fn run_program_default(program: &Program) -> Result<ProgramResult> {
-        let mut ctx = VmContext::new();
-        run_program(program, &mut ctx)
-    }
-}
