+++ stdlib/crates/iter/src/lib.rs
@@ -906,9 +906,10 @@ fn runtime_value_equals_string(value: &RuntimeVal, expected: &str, heap: &HeapSt
 #[cfg(test)]
 mod tests {
     use super::*;
+    use lk_core::vm::ModuleResolver;
     use lk_core::vm::ProgramExec;
     use lk_core::{
-        stmt::{ModuleResolver, stmt_parser::StmtParser},
+        stmt::stmt_parser::StmtParser,
         token::Tokenizer,
         vm::{NativeFunction, ProgramResult, RuntimeModuleState, VmContext},
     };
