+++ stdlib/src/os_test.rs
@@ -1,5 +1,6 @@
 #[cfg(test)]
 mod tests {
+    use lk_core::vm::ModuleResolver;
     use lk_core::vm::ProgramExec;
     use std::sync::Arc;
 
@@ -7,7 +8,7 @@ mod tests {
     use anyhow::Result;
     use lk_core::{
         module::ModuleRegistry,
-        stmt::{ModuleResolver, stmt_parser::StmtParser},
+        stmt::stmt_parser::StmtParser,
         token::Tokenizer,
         val::{HeapStore, HeapValue, RuntimeVal},
         vm::{NativeFunction, ProgramResult, VmContext},
