+++ stdlib/src/select_test.rs
@@ -6,13 +6,14 @@
 
 #[cfg(test)]
 mod tests {
+    use lk_core::vm::ModuleResolver;
     use lk_core::vm::ProgramExec;
     use std::sync::Arc;
 
     use anyhow::Result;
     use lk_core::{
         module::ModuleRegistry,
-        stmt::{ModuleResolver, stmt_parser::StmtParser},
+        stmt::stmt_parser::StmtParser,
         token::Tokenizer,
         val::RuntimeVal,
         vm::{ProgramResult, VmContext},
