+++ stdlib/src/stdlib_modules_test.rs
@@ -1,12 +1,13 @@
 #[cfg(test)]
 mod tests {
+    use lk_core::vm::ModuleResolver;
     use lk_core::vm::ProgramExec;
     use std::{fs::File, io::Write, sync::Arc};
 
     use anyhow::Result;
     use lk_core::{
         module::ModuleRegistry,
-        stmt::{ModuleResolver, stmt_parser::StmtParser},
+        stmt::stmt_parser::StmtParser,
         token::Tokenizer,
         val::{CallableValue, HeapValue, RuntimeVal, TypedList},
         vm::{ProgramResult, VmContext},
