+++ stdlib/src/stream_test.rs
@@ -1,9 +1,10 @@
 #[cfg(test)]
 mod tests {
     use anyhow::Result;
+    use lk_core::vm::ModuleResolver;
     use lk_core::vm::ProgramExec;
     use lk_core::{
-        stmt::{ModuleResolver, stmt_parser::StmtParser},
+        stmt::stmt_parser::StmtParser,
         token::Tokenizer,
         val::{HeapStore, HeapValue, RuntimeVal, TypedList},
         vm::{NativeArgs, NativeFunction, NativeRuntime, ProgramResult, RuntimeModuleState, VmContext},
